+++ changelog.d/pgw1215.md
@@ -1,0 +1,28 @@
+- **pgw#1215 (th#1834 Phase 3): the mint's shared substrate leaves the modules that are about to
+  die.** Phase 3 replaces the out-of-process mint driver with a serving parent that supervises
+  compile children directly, one graph class each — at which point `mint_child`, `mint_process` and
+  `mint_delegate` fall out. Measured first: **they are ~20 % substrate, not driver.** `mint_process`
+  is also the tree's home for the parent↔child *vocabulary* and `mint_child` for the child-side
+  *preflight*, and five modules that SURVIVE the rewrite already import both — `executor`,
+  `boot_key`, `boot_adopt`, `measure_child`, `local_serve` — plus `boot_trace_child` and both
+  operator rigs.
+
+  So the substrate moves out first, on its own, with no behaviour change: `child_contract.py`
+  (`MintSlot`, `CompileSpec`, `slot_subjects`, `MintFrame`, `frame_line`) and `child_preflight.py`
+  (`PreflightRefused`, `select_specs`, `bind_slots`, `assert_slots_resolvable`,
+  `pick_compile_target`). With it gone the driver's removal is a **cut**; with it still inside, that
+  removal would have been an untangle — a deletion obliged to preserve two thirds of what it deletes,
+  which is how a rewrite acquires a compatibility shim nobody asked for.
+
+  **A defect fixed on the way out.** `boot_trace_child` and `measure_child` reached into `mint_child`
+  for the same five preflight symbols through **function-local imports at three call sites**, because
+  importing `mint_child` at module scope drags the whole compile brain in behind it — the shape §1.17
+  forbids, and unfixable while the checks lived beside the driver. They are ordinary top-of-file
+  imports now, and the module that owns the mint has stopped being a library for children that do not
+  mint.
+
+  **Renames** (no aliases, pre-launch): `CompileCellSpec` → `CompileSpec` and `MintChildRefused` →
+  `PreflightRefused`. The refusal is deliberately NOT called `MintRefused` — that name is live in
+  `aot_contract` for "every declared graph class refused", and one name for both would make "this
+  subject cannot be built here" and "nothing compiled" indistinguishable at the exit-code boundary
+  with no test going red.
