+++ src/gen_worker/mint_child.py
@@ -67,6 +67,14 @@ from . import compile_posture, handler_proof, warm_spans
 from .api.errors import ValidationError
 from .api.export_contract import (
     blocker_refusal, export_declaration, open_blockers)
+from .child_contract import MintSlot, frame_line
+from .child_preflight import (
+    PreflightRefused,
+    assert_slots_resolvable,
+    bind_slots,
+    pick_compile_target,
+    select_specs,
+)
 from .mint_process import (
     EXIT_BAD_REQUEST,
     EXIT_MINTED,
@@ -74,8 +82,6 @@ from .mint_process import (
     EXIT_RESOURCE,
     MintReport,
     MintRequest,
-    MintSlot,
-    frame_line,
 )
 
 logger = logging.getLogger(__name__)
@@ -84,24 +90,6 @@ logger = logging.getLogger(__name__)
 #: also run produced a cell with no consumer, and it is deleted rather than
 #: kept behind a request field nobody may set.
 RECIPE_AOT = "aot"
-
-
-class MintChildRefused(RuntimeError):
-    """A named, deterministic reason this mint cannot happen here.
-
-    Never retried by the parent: re-running a named refusal buys a second
-    billed compile for the same sentence.
-
-    ``mint_phases`` carries the refusing mint's PARTIAL phase table when it
-    got far enough to have one (pgw#825) — the entries it exported and
-    compiled before refusing are real minutes and must reach the hub.
-    """
-
-    def __init__(
-        self, *args: Any, mint_phases: Optional[Mapping[str, Any]] = None,
-    ) -> None:
-        super().__init__(*args)
-        self.mint_phases: Dict[str, Any] = dict(mint_phases or {})
 
 
 def _assert_family_mintable(family: str) -> None:
@@ -119,17 +107,17 @@ def _assert_family_mintable(family: str) -> None:
     try:
         decl = export_declaration(family)
     except Exception as exc:  # noqa: BLE001 — a refusing declaration is a refusal
-        raise MintChildRefused(
+        raise PreflightRefused(
             f"family {family!r}'s export declaration refuses to build "
             f"({type(exc).__name__}): {exc}") from exc
     if decl is None:
         return
     blocked = open_blockers(decl)
     if blocked:
-        raise MintChildRefused(blocker_refusal(family, blocked))
+        raise PreflightRefused(blocker_refusal(family, blocked))
 
 
-def _declaration_refusal(exc: ValidationError) -> MintChildRefused:
+def _declaration_refusal(exc: ValidationError) -> PreflightRefused:
     """pgw#1075: a declared value this SDK's own vocabulary rejects is a
     REFUSAL, not a crash — and the author's fix is already in the message.
 
@@ -153,7 +141,7 @@ def _declaration_refusal(exc: ValidationError) -> MintChildRefused:
     to a custom op with no fake kernel (pgw#1062) — the message is the
     authoring contract, so it is carried whole and never summarised.
     """
-    return MintChildRefused(
+    return PreflightRefused(
         f"declaration refused: {type(exc).__name__}: {exc}",
         mint_phases=getattr(exc, "mint_phases", None))
 
@@ -207,34 +195,6 @@ def _write_report(path: Path, report: MintReport) -> None:
     os.replace(tmp, path)
 
 
-def select_specs(
-    specs: Sequence[Any], function: str,
-) -> Tuple[Any, List[Any]]:
-    """The requested function's spec plus every sibling sharing its instance.
-
-    Siblings matter: the warm plan is CLASS-scoped (pgw#654 — one cell per
-    class, so turbo and base seed the same capture). Minting from one
-    function's view is how a sibling lane ends up absent from the cell and
-    every adopting boot fails its per-object proof (gw#612).
-    """
-    chosen = next((s for s in specs if s.name == function), None)
-    if chosen is None:
-        raise MintChildRefused(
-            f"function {function!r} is not in this image "
-            f"(have: {sorted(s.name for s in specs)})")
-    if chosen.cls is None:
-        raise MintChildRefused(
-            f"function {function!r} is function-shaped — it has no instance "
-            "to compile")
-    siblings = [
-        s for s in specs
-        if s.cls is chosen.cls and s.instance_key == chosen.instance_key
-    ]
-    if chosen not in siblings:
-        siblings.append(chosen)
-    return chosen, siblings
-
-
 def mint_identity(request: MintRequest) -> str:
     """One sentence naming WHICH mint a message belongs to (pgw#969).
 
@@ -248,88 +208,6 @@ def mint_identity(request: MintRequest) -> str:
         f"mint family={request.family!r} arm_key={request.arm_token!r} "
         f"lane={request.execution_lane or '(unset)'!r} "
         f"fn={request.function!r}")
-
-
-def bind_slots(specs: Sequence[Any], resolved: Mapping[str, MintSlot]) -> None:
-    """Install the PARENT's resolved slot bindings onto rediscovered specs.
-
-    The child re-runs discovery in a fresh interpreter, so ``spec.models``
-    comes back holding only what ``@endpoint`` declared. For a hub-catalog
-    slot — ``Slot(cls, selected_by="model")`` with no ``default_checkpoint=``,
-    which is sdxl's shape and every multi-checkpoint family's — the decorator
-    declares no ref at all, and the resolution chain then has nothing to
-    resolve: ``ctx.slots["pipeline"]`` raises before a graph is traced.
-
-    Applied to the WHOLE sibling set, never one spec: ``instance_key`` is a
-    live property over ``spec.models``, so binding the chosen function alone
-    would move its key out from under its siblings and silently narrow the
-    class-scoped warm plan (pgw#654) to one lane.
-    """
-    for spec in specs:
-        for slot, res in resolved.items():
-            if slot in spec.slots or slot in spec.models:
-                spec.models[slot] = res.ref
-
-
-def assert_slots_resolvable(
-    specs: Sequence[Any],
-    slots: Mapping[str, MintSlot],
-    *,
-    what: str = "",
-) -> None:
-    """pgw#969: refuse a request whose slots cannot resolve — before the load.
-
-    ``slots`` + ``what`` rather than the whole ``MintRequest`` (pgw#1089): the
-    boot-trace child resolves the same slots for the same reason and must get
-    the same refusal, and a second copy of this check would be a second answer
-    to "can this process trace the checkpoint the parent serves".
-
-    Two distinct failures, both named rather than deferred to a handler's
-    first dereference nine seconds later:
-
-    * the parent sent no binding for a declared, non-optional slot (the wire
-      gap this issue exists for), and
-    * a slot the parent DID bind still fails the resolution chain here, which
-      is real divergence between the two processes and never a normal path.
-
-    An OPTIONAL slot the parent did not bind is deliberately silent: the
-    parent's own warm context resolves it exactly this way (the deploy chose
-    not to serve that lane), and refusing here would refuse a mint the parent
-    can serve.
-    """
-    from . import warmup as warmup_mod
-
-    bound = set(slots)
-    problems: List[str] = []
-    for spec in specs:
-        missing = sorted(
-            name for name, slot in spec.slots.items()
-            if name not in bound
-            and name not in spec.models
-            and not getattr(slot, "optional", False)
-        )
-        for name in missing:
-            problems.append(
-                f"slot {name!r} of {spec.name!r}: the parent sent no resolved "
-                f"binding for it and the endpoint declares no "
-                f"default_checkpoint, so this child has no checkpoint to "
-                f"trace (MintRequest.slots carries "
-                f"{sorted(bound) or '(nothing)'})")
-        errors = warmup_mod.resolved_slots_kwargs(spec, None)["slot_errors"]
-        for name, why in sorted(errors.items()):
-            resolved = slots.get(name)
-            if resolved is not None:
-                problems.append(
-                    f"slot {name!r} of {spec.name!r}: the parent's binding "
-                    f"{resolved.ref.path!r} "
-                    f"crossed but does not resolve in this process — {why}")
-    if not problems:
-        return
-    raise MintChildRefused(
-        f"{what or 'slot resolution'}: cannot build the warmup request — "
-        + "; ".join(problems)
-        + ". A mint cannot trace a graph for a pipeline it was never told "
-          "to load.")
 
 
 def assert_composable(resolved: Mapping[str, MintSlot]) -> None:
@@ -355,32 +233,12 @@ def assert_composable(resolved: Mapping[str, MintSlot]) -> None:
     )
     if not bad:
         return
-    raise MintChildRefused(
+    raise PreflightRefused(
         f"slot(s) {bad} were materialized as override-narrowed trees "
         f"(the overridden component's files were excluded from the fetch) "
         f"but this request carries no component override for them, so the "
         f"composition cannot be rebuilt: "
         + "; ".join(f"{slot}={resolved[slot].path}" for slot in bad))
-
-
-def pick_compile_target(loaded: Dict[str, Any], cfg: Any) -> Tuple[str, Any]:
-    """The loaded slot that actually carries the declared compile target(s).
-
-    Public since pgw#1089: the boot-trace child must pick the SAME slot this
-    child picks, and two pickers would be two compile targets.
-    """
-    from . import compile_cache as cc
-
-    for slot, obj in loaded.items():
-        try:
-            if cc.has_compile_target(obj, cfg):
-                return slot, obj
-        except Exception:
-            continue
-    raise MintChildRefused(
-        f"no compile target resolved on any loaded slot "
-        f"({sorted(loaded) or '(none)'}) for family {cfg.family!r} "
-        f"targets={list(cfg.targets)}")
 
 
 def _drive_warm_plan(
@@ -423,7 +281,7 @@ def _drive_warm_plan(
         except BaseException as exc:
             if _is_resource_error(exc) or not isinstance(exc, Exception):
                 raise
-            raise MintChildRefused(
+            raise PreflightRefused(
                 f"{mint_identity(request)}: the endpoint's own warm plan does "
                 f"not run — warm job {job.spec.name!r} raised "
                 f"{type(exc).__name__}: {exc}. A cell must not seal for a "
@@ -637,7 +495,7 @@ def _mint_aot(
         # that exposes no logs (pgw#760). pgw#825: the seconds it spent before
         # refusing ride WITH the sentence — a refusal after four paid compiles
         # and a refusal in the first second are not the same event.
-        raise MintChildRefused(
+        raise PreflightRefused(
             f"aot mint refused: {exc}",
             mint_phases=getattr(exc, "mint_phases", None)) from exc
 
@@ -697,7 +555,7 @@ def _mint_aot(
 
 
 def mint(request: MintRequest) -> MintReport:
-    """Build the cell. Raises ``MintChildRefused`` for a named refusal."""
+    """Build the cell. Raises ``PreflightRefused`` for a named refusal."""
     # The two views ``cli.run.run_setup`` takes, both DERIVED from the one
     # resolution (pgw#974) rather than carried beside it.
     paths = {slot: res.path for slot, res in request.slots.items()}
@@ -723,14 +581,14 @@ def mint(request: MintRequest) -> MintReport:
     env_seal.establish()
 
     if not cc.toolchain_present():
-        raise MintChildRefused(
+        raise PreflightRefused(
             "no C toolchain (cc/gcc/clang) — inductor cannot link a kernel")
     if not cc.cxx_toolchain_present():
         # pgw#823: AOTInductor's stricter requirement, asserted BEFORE the
         # weights are read. The guard above passes on a C compiler; AOTI needs
         # a C++ one, and without this the miss surfaces 336 s later as an
         # InductorError from inside the linker.
-        raise MintChildRefused(
+        raise PreflightRefused(
             "no C++ compiler (g++/clang++) — AOTInductor links a shared "
             "object and cannot build one")
 
@@ -824,7 +682,7 @@ def mint(request: MintRequest) -> MintReport:
             # (ie#638). This is buildable-but-not-honored, NOT a stranded
             # family, so it FAILS CLOSED with the authoring cause named rather
             # than degrading to a real-weight export the meta gate never sees.
-            raise MintChildRefused(
+            raise PreflightRefused(
                 f"structure-only was requested for {mint_identity(request)} "
                 f"and the target built weight-free, but the composed pipeline "
                 f"did not carry it: {exc}") from exc
@@ -972,7 +830,7 @@ def assert_traceable_as_loaded(pipeline: Any, request: MintRequest) -> None:
     reason = traceability.untraceable_reason(pipeline)
     if not reason:
         return
-    raise MintChildRefused(
+    raise PreflightRefused(
         f"mint_pipeline_not_traceable: {mint_identity(request)} loaded real "
         f"weights and the placement ladder installed offload hooks — {reason}. "
         f"`torch.export` refuses a `torch.compiler.disable`d function rather "
@@ -1009,7 +867,7 @@ def assert_handler_proven(request: MintRequest) -> None:
     """
     if str(getattr(request, "handler_proof", "") or "").strip():
         return
-    raise MintChildRefused(
+    raise PreflightRefused(
         f"{mint_identity(request)}: this is a WEIGHT-FREE mint and the parent "
         f"sent no handler proof. pgw#984 requires the endpoint's own handler "
         f"to have run before a cell seals, and since pgw#1199 that proof "
@@ -1178,7 +1036,7 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
             # this process — spec build, composition check, branch arm, export
             # declaration — can raise it and they all mean the same thing.
             raise _declaration_refusal(exc) from exc
-    except MintChildRefused as exc:
+    except PreflightRefused as exc:
         # th#1322: a refused mint's phase table is where it spent the time
         # BEFORE refusing — the most useful half of a failed mint. The OPEN
         # phase is read first on purpose: `_close_phases` closes it, and
@@ -1217,7 +1075,4 @@ if __name__ == "__main__":  # pragma: no cover
     raise SystemExit(main())
 
 
-__all__ = ["MintChildRefused", "assert_composable", "assert_slots_resolvable",
-           "pick_compile_target",
-           "bind_slots", "frame", "main",
-           "mint_identity", "mint", "select_specs"]
+__all__ = ["assert_composable", "frame", "main", "mint_identity", "mint"]
