+++ tests/test_measure_document_pgw1153.py
@@ -45,7 +45,8 @@ import msgspec
 import pytest
 
 from gen_worker import measure_child
-from gen_worker.mint_process import CompileCellSpec, MintRequest
+from gen_worker.child_contract import CompileSpec
+from gen_worker.mint_process import MintRequest
 
 REPO = Path(__file__).resolve().parent.parent
 MICRO_SRC = REPO / "examples" / "micro-diffusion" / "src"
@@ -155,7 +156,7 @@ def test_a_runtime_mint_request_still_decodes_through_the_same_door(
 ) -> None:
     """The OTHER real artifact — a mint work root's ``request.json`` — has to
     keep working, and through the SAME decoder. Two decoders is the defect."""
-    cfg = CompileCellSpec(family=FAMILY, targets=("transformer",))
+    cfg = CompileSpec(family=FAMILY, targets=("transformer",))
     request = MintRequest(
         function="generate-w8a8", modules=("micro_diffusion.main_w8a8",),
         family=FAMILY, arm_token="arm1-deadbeef",
