+++ tests/test_mint_preconditions_pgw996.py
@@ -22,6 +22,7 @@ from pathlib import Path
 
 import pytest
 
+from gen_worker import child_preflight
 from gen_worker import aot_preconditions as pre
 from gen_worker import compile_cache as cc
 from gen_worker import fleet_cells
@@ -409,7 +410,7 @@ def test_the_CHILD_still_refuses_a_toolchainless_AOT_mint(
         target="/tmp/x", capture="/tmp/y", device=0,
         modules=[], function="generate", cfg=None, configs={},
         execution_lane="", report="/tmp/r", arm_token="")
-    with pytest.raises(mint_child.MintChildRefused, match="no C\\+\\+ compiler"):
+    with pytest.raises(child_preflight.PreflightRefused, match="no C\\+\\+ compiler"):
         mint_child.mint(req)
